+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
+
+## [0.1.7][] - 2017-07-19
 ### Fixed
 - Header changing size from CSS causing the background to not update their position
 
@@ -46,7 +48,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Things that have changed
 
 
-[Unreleased]: https://github.com/madou/react-sticky-header/compare/v0.1.6...HEAD
+[Unreleased]: https://github.com/madou/react-sticky-header/compare/v0.1.7...HEAD
+[0.1.7]: https://github.com/madou/react-sticky-header/compare/v0.1.6...v0.1.7
 [0.1.6]: https://github.com/madou/react-sticky-header/compare/v0.1.5...v0.1.6
 [0.1.5]: https://github.com/madou/react-sticky-header/compare/v0.1.4...v0.1.5
 [0.1.4]: https://github.com/madou/react-sticky-header/compare/v0.1.3...v0.1.4